requirements.txt
# 数控车床多订单宏变量管理与生产任务调度系统依赖包

# GUI框架
PyQt6==6.5.2
PySide6==6.5.2

# UI自动化
pywinauto==0.6.8
selenium==4.11.2
webdriver-manager==3.8.6

# 文件监控
watchdog==3.0.0

# 数据处理
pandas==2.0.3
numpy==1.24.3

# 配置管理
PyYAML==6.0
configparser==5.3.0

# 数据库
sqlite3

# 图像处理
opencv-python==4.8.0.74
Pillow==10.0.0

# 网络请求
requests==2.31.0

# 日志管理
loguru==0.7.2

# 其他工具
python-dateutil==2.8.2
typing-extensions==4.7.1
